[WFCORE-6206] Do not enable the installer high level commands for embedded servers or when the installer resource is not available
