+++ installation-manager/src/main/java/org/wildfly/core/instmgr/cli/CustomPatchUploadCommand.java
@@ -17,7 +17,8 @@ import org.wildfly.core.cli.command.aesh.CLICommandInvocation;
 import org.wildfly.core.instmgr.InstMgrConstants;
 import org.wildfly.core.instmgr.InstMgrCustomPatchUploadHandler;
 
-@CommandDefinition(name = "upload-custom-patch", description = "Uploads a custom patch Zip file to the server and subscribes the installation creating a channel to handle the patch.")
+@CommandDefinition(name = "upload-custom-patch", description = "Uploads a custom patch Zip file to the server and subscribes the installation creating a channel to handle the patch.",
+        activator = InstMgrActivator.class)
 public class CustomPatchUploadCommand extends AbstractInstMgrCommand {
 
     @Option(name = "custom-patch-file", required = true)
